add method to add new limit

+++ src/main/java/com/github/pavelvashkevich/bankmicroservice/service/BankAccountService.java
@@ -4,7 +4,7 @@ import com.github.pavelvashkevich.bankmicroservice.model.postgres.BankAccount;
 
 public interface BankAccountService {
 
-    BankAccount findByAccountNumber(int accountNumber);
+    BankAccount findByAccountNumber(long accountNumber);
 
     boolean isBankAccountWithNumberExist(long accountNumber);
 }
